+++ src/backend/ServicesDeskUCABWS/BussinesLogic/DAO/TipoEstadoDAO/TipoEstadoService.cs
@@ -26,11 +26,6 @@ namespace ServicesDeskUCABWS.BussinesLogic.DAO.TipoEstadoDAO
             _estadoService = estadoService;
         }
 
-        public TipoEstadoService(IDataContext tipoestadoContext)
-        {
-            _tipoEstadoContext = tipoestadoContext;
-        }
-
         //GET: Servicio para consultar todos los tipos estados
         public List<TipoEstadoDTO> ConsultaTipoEstados()
         {
@@ -123,8 +118,8 @@ namespace ServicesDeskUCABWS.BussinesLogic.DAO.TipoEstadoDAO
 
                 _tipoEstadoContext.Tipos_Estados.Add(tipoEstadoEntity);
 
-                //Llena la entidad intermedia Estado
-                AgregarEstadoATipoEstadoCreado(tipoEstadoEntity);
+				//Llena la entidad intermedia Estado
+				_estadoService.AgregarEstadoATipoEstadoCreado(tipoEstadoEntity);
 
                 _tipoEstadoContext.DbContext.SaveChanges();
                 return _mapper.Map<TipoEstadoDTO>(tipoEstadoEntity);
@@ -141,27 +136,6 @@ namespace ServicesDeskUCABWS.BussinesLogic.DAO.TipoEstadoDAO
             {
                 throw new ExceptionsControl("No se pudo registrar el tipo estado", ex);
             }            
-        }
-
-        //Agregar Estados de los Tipo Estados Agregados
-        public void AgregarEstadoATipoEstadoCreado(Tipo_Estado estado)
-        {
-            var listaEstados = new List<Estado>();
-
-            foreach (var departamento in _tipoEstadoContext.Departamentos.ToList())
-            {
-                listaEstados.Add(new Estado(departamento.nombre + " " + estado.nombre, estado.descripcion)
-                {
-                    Id = Guid.NewGuid(),
-                    Departamento = departamento,
-                    Estado_Padre = estado,
-                    Bitacora_Tickets = new List<Bitacora_Ticket>(),
-                    ListaTickets = new List<Ticket>()
-                });
-            }
-
-            _tipoEstadoContext.Estados.AddRange(listaEstados);
-            _tipoEstadoContext.DbContext.SaveChanges();
         }
 
         //PUT: Servicio para actualizar tipo estado
